fix: Redis reconnection with ElastiCache/MemoryDB Failover handling
