+++ Csharp-lab.sln
@@ -126,10 +126,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CSharp10.Tests", "src\Whats
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CSharp6.Tests", "src\WhatsNew\CSharp6.Tests\CSharp6.Tests.csproj", "{6CA0D0AB-721D-401E-8592-8B2F81DF7B3B}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RedisFailover", "src\Redis\RedisFailover\RedisFailover.csproj", "{473A75DE-21AD-4A1C-BF50-59D3D1D4D1E2}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RedisFailoverDirect", "src\Redis\RedisFailoverDirect\RedisFailoverDirect.csproj", "{BCD1E0A6-9E62-4B22-86FF-F1207313615A}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SystemInfo", "SystemInfo", "{758C0491-77D7-45B6-9E03-5EE26671D447}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SystemInfo.Core", "src\SystemInfo\SystemInfo.Core\SystemInfo.Core.csproj", "{9BFE753D-C2D9-4022-A990-70A9831F6CDD}"
@@ -145,6 +141,10 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ApiGrpcHttp2", "src\Api\ApiGrpcHttp2\ApiGrpcHttp2.csproj", "{D0C2D015-2DC6-4BAC-912A-3E394EF76895}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ApiGrpcHttp3", "src\Api\ApiGrpcHttp3\ApiGrpcHttp3.csproj", "{2A01FF0C-BCAE-4605-AA03-DB84AF7BA934}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RedisFailover.Managed", "src\Redis\RedisFailover.Managed\RedisFailover.Managed.csproj", "{9578D3E5-9E4B-4ABA-91E9-579EAE967733}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RedisFailover.Direct", "src\Redis\RedisFailover.Direct\RedisFailover.Direct.csproj", "{99B26CD5-E8EF-430E-B0EA-DFDAAE9149DA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -324,14 +324,6 @@ Global
 		{6CA0D0AB-721D-401E-8592-8B2F81DF7B3B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6CA0D0AB-721D-401E-8592-8B2F81DF7B3B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6CA0D0AB-721D-401E-8592-8B2F81DF7B3B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{473A75DE-21AD-4A1C-BF50-59D3D1D4D1E2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{473A75DE-21AD-4A1C-BF50-59D3D1D4D1E2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{473A75DE-21AD-4A1C-BF50-59D3D1D4D1E2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{473A75DE-21AD-4A1C-BF50-59D3D1D4D1E2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{BCD1E0A6-9E62-4B22-86FF-F1207313615A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BCD1E0A6-9E62-4B22-86FF-F1207313615A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BCD1E0A6-9E62-4B22-86FF-F1207313615A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BCD1E0A6-9E62-4B22-86FF-F1207313615A}.Release|Any CPU.Build.0 = Release|Any CPU
 		{9BFE753D-C2D9-4022-A990-70A9831F6CDD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{9BFE753D-C2D9-4022-A990-70A9831F6CDD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9BFE753D-C2D9-4022-A990-70A9831F6CDD}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -360,6 +352,14 @@ Global
 		{2A01FF0C-BCAE-4605-AA03-DB84AF7BA934}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2A01FF0C-BCAE-4605-AA03-DB84AF7BA934}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2A01FF0C-BCAE-4605-AA03-DB84AF7BA934}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9578D3E5-9E4B-4ABA-91E9-579EAE967733}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9578D3E5-9E4B-4ABA-91E9-579EAE967733}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9578D3E5-9E4B-4ABA-91E9-579EAE967733}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9578D3E5-9E4B-4ABA-91E9-579EAE967733}.Release|Any CPU.Build.0 = Release|Any CPU
+		{99B26CD5-E8EF-430E-B0EA-DFDAAE9149DA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{99B26CD5-E8EF-430E-B0EA-DFDAAE9149DA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{99B26CD5-E8EF-430E-B0EA-DFDAAE9149DA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{99B26CD5-E8EF-430E-B0EA-DFDAAE9149DA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -408,8 +408,6 @@ Global
 		{0CC5D10F-9792-4E95-9CB5-6244979D5A4A} = {F611191C-1089-4855-A8AA-ECE6F4586743}
 		{F4E57CF9-A1E9-46D8-84CF-354C962377B5} = {D2828C01-951E-4344-B225-DBCDB9502F6A}
 		{6CA0D0AB-721D-401E-8592-8B2F81DF7B3B} = {D2828C01-951E-4344-B225-DBCDB9502F6A}
-		{473A75DE-21AD-4A1C-BF50-59D3D1D4D1E2} = {3910F58C-6DEB-4C30-8B10-B9DA74B58225}
-		{BCD1E0A6-9E62-4B22-86FF-F1207313615A} = {3910F58C-6DEB-4C30-8B10-B9DA74B58225}
 		{9BFE753D-C2D9-4022-A990-70A9831F6CDD} = {758C0491-77D7-45B6-9E03-5EE26671D447}
 		{99FF245D-EB68-4623-B586-8E37B8E2D8FD} = {758C0491-77D7-45B6-9E03-5EE26671D447}
 		{992BD9EA-768B-4A51-86E6-3B661F6D18A8} = {28CE7184-F1DE-44AE-922F-965FC3AA1768}
@@ -417,6 +415,8 @@ Global
 		{EE9E851C-8E09-4C04-B7AE-AE165353F5FA} = {28CE7184-F1DE-44AE-922F-965FC3AA1768}
 		{D0C2D015-2DC6-4BAC-912A-3E394EF76895} = {28CE7184-F1DE-44AE-922F-965FC3AA1768}
 		{2A01FF0C-BCAE-4605-AA03-DB84AF7BA934} = {28CE7184-F1DE-44AE-922F-965FC3AA1768}
+		{9578D3E5-9E4B-4ABA-91E9-579EAE967733} = {3910F58C-6DEB-4C30-8B10-B9DA74B58225}
+		{99B26CD5-E8EF-430E-B0EA-DFDAAE9149DA} = {3910F58C-6DEB-4C30-8B10-B9DA74B58225}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {6AD43C70-ADE6-43C4-A90D-F4E7CB0BD2F9}
