+++ src/Redis/RedisFailover.Direct/Properties/launchSettings.json
