+++ src/Redis/RedisFailover.Direct/RedisFailover.Direct.csproj
