+++ src/Redis/RedisFailover.Direct/appsettings.Development.json
