+++ src/Redis/RedisFailover.Direct/appsettings.json
