+++ src/Redis/RedisFailover.Managed/Dockerfile
@@ -2,7 +2,7 @@
 
 # This stage is used when running from VS in fast mode (Default for Debug configuration)
 FROM mcr.microsoft.com/dotnet/aspnet:9.0 AS base
-USER app
+USER $APP_UID
 WORKDIR /app
 EXPOSE 8080
 
@@ -11,20 +11,21 @@ EXPOSE 8080
 FROM mcr.microsoft.com/dotnet/sdk:9.0 AS build
 ARG BUILD_CONFIGURATION=Release
 WORKDIR /src
-COPY ["Directory.Build.props", "src/Directory.Build.props"]
-COPY ["src/Redis/RedisFailover/RedisFailover.csproj", "src/Redis/RedisFailover/"]
-RUN dotnet restore "./src/Redis/RedisFailover/RedisFailover.csproj"
+COPY ["Directory.Packages.props", "."]
+COPY ["Directory.Build.props", "."]
+COPY ["src/Redis/RedisFailover.Managed/RedisFailover.Managed.csproj", "src/Redis/RedisFailover.Managed/"]
+RUN dotnet restore "./src/Redis/RedisFailover.Managed/RedisFailover.Managed.csproj"
 COPY . .
-WORKDIR "/src/src/Redis/RedisFailover"
-RUN dotnet build "./RedisFailover.csproj" -c $BUILD_CONFIGURATION -o /app/build
+WORKDIR "/src/src/Redis/RedisFailover.Managed"
+RUN dotnet build "./RedisFailover.Managed.csproj" -c $BUILD_CONFIGURATION -o /app/build
 
 # This stage is used to publish the service project to be copied to the final stage
 FROM build AS publish
 ARG BUILD_CONFIGURATION=Release
-RUN dotnet publish "./RedisFailover.csproj" -c $BUILD_CONFIGURATION -o /app/publish /p:UseAppHost=false
+RUN dotnet publish "./RedisFailover.Managed.csproj" -c $BUILD_CONFIGURATION -o /app/publish /p:UseAppHost=false
 
 # This stage is used in production or when running from VS in regular mode (Default when not using the Debug configuration)
 FROM base AS final
 WORKDIR /app
 COPY --from=publish /app/publish .
-ENTRYPOINT ["dotnet", "RedisFailover.dll"]
+ENTRYPOINT ["dotnet", "RedisFailover.Managed.dll"]
