+++ src/Redis/RedisFailover.Managed/Properties/launchSettings.json
