+++ src/Redis/RedisFailover.Managed/RedisFailover.Managed.csproj
