+++ src/Redis/RedisFailover.Managed/RedisFailover.http
