+++ src/Redis/RedisFailover.Managed/appsettings.Development.json
