+++ src/Redis/RedisFailover.Managed/appsettings.json
